+++ core/dbt/include/global_project/macros/adapters/common.sql
@@ -65,9 +65,6 @@
 {% endmacro %}
 
 {% macro create_table_as(temporary, relation, sql) -%}
-  {%- if not temporary -%}
-    {%- do adapter.cache_added(relation.incorporate(dbt_created=True)) -%}
-  {%- endif -%}
   {{ adapter_macro('create_table_as', temporary, relation, sql) }}
 {%- endmacro %}
 
@@ -80,7 +77,6 @@
 {% endmacro %}
 
 {% macro create_view_as(relation, sql) -%}
-  {%- do adapter.cache_added(relation.incorporate(dbt_created=True)) -%}
   {{ adapter_macro('create_view_as', relation, sql) }}
 {%- endmacro %}