+++ core/dbt/include/global_project/macros/materializations/incremental/incremental.sql
@@ -79,4 +79,6 @@
 
   {{ run_hooks(post_hooks, inside_transaction=False) }}
 
+  {{ return({'relations': [target_relation]}) }}
+
 {%- endmaterialization %}