+++ core/dbt/include/global_project/macros/materializations/seed/seed.sql
@@ -1,10 +1,5 @@
 
 {% macro create_csv_table(model, agate_table) -%}
-  {%- do adapter.cache_added(api.Relation.create(
-    database=model.database,
-    schema=model.schema,
-    identifier=model.alias,
-    dbt_created=True)) -%}
   {{ adapter_macro('create_csv_table', model, agate_table) }}
 {%- endmacro %}
 
@@ -136,4 +131,8 @@
   {{ adapter.commit() }}
 
   {{ run_hooks(post_hooks, inside_transaction=False) }}
+
+  {% set target_relation = this.incorporate(type='table') %}
+  {{ return({'relations': [target_relation]}) }}
+
 {% endmaterialization %}